assignment 2 - modified
This one, I used the NSString stringWithFormat in all my calls within
the detail view controller to add in the descriptors. Somethings I
noticed was that it is hard to judge the amount of space being used by
the labels. I had to run it and adjust it a few times before the
spacing was right. Also, I think I didn't  us the navigationHeader.txt
call correctly, because my restaurant name did not show up. I manual
input it within storyboard mode.